refactor: unify poll UI
