+++ src/app/poll/[pollId]/page.tsx
@@ -6,34 +6,14 @@ import { useEffect } from "react";
 import { useQuery } from "convex/react";
 import { useMiniApp } from "@neynar/react";
 import { useParams, notFound } from "next/navigation";
-import { CheckCircle, Users, Badge, Share2 } from "lucide-react";
-import { Card, CardContent, CardHeader, CardTitle } from "~/components/ui/card";
-import { Button } from "~/components/ui/Button";
-import { ShareButton } from "~/components/ui/Share";
-import Link from "next/link";
+import { PollCard } from "~/components/poll/PollCard";
 import { api } from "../../../../convex/_generated/api";
 import { Id } from "../../../../convex/_generated/dataModel";
-import { APP_URL } from "~/lib/constants";
+import Link from "next/link";
 
 type PollPageParams = {
   pollId: string;
 };
-
-interface PollOption {
-  id: string;
-  text: string;
-  votes: number;
-}
-
-interface Poll {
-  _id: Id<"polls">;
-  title: string;
-  description: string;
-  options: PollOption[];
-  totalVotes: number;
-  creatorFid: string;
-  createdAt: number;
-}
 
 
 
@@ -44,10 +24,6 @@ export default function PollPage() {
 
   // Fetch poll and vote status using useQuery
   const poll = useQuery(api.polls.getPoll, pollId ? { pollId: pollId as Id<"polls"> } : "skip");
-  const voteStatus = useQuery(
-    api.polls.hasVoted,
-    userFid && pollId ? { pollId: pollId as Id<"polls">, userFid } : "skip"
-  );
 
   // Defer notFound() to useEffect to avoid render-time side effects
   useEffect(() => {
@@ -71,21 +47,6 @@ export default function PollPage() {
     return null; // This won't be reached due to useEffect, but kept for type safety
   }
 
-  const castConfig = {
-    text: `Check out this poll: ${poll.title || "Untitled"} on Flash Poll! @1 @2`,
-    bestFriends: true,
-    embeds: [
-      {
-        path: `/poll/${pollId}`,
-        imageUrl: async () => `${APP_URL}/api/opengraph-image?pollId=${pollId}`,
-      },
-    ],
-  };
-
-  const getPercentage = (votes: number, total: number): number => {
-    return total > 0 ? Math.round((votes / total) * 100) : 0;
-  };
-
   return (
     <div className="min-h-screen bg-gradient-to-br from-purple-50 via-white to-blue-50">
       <div className="bg-white/70 backdrop-blur-md border-b border-purple-100 sticky top-0 z-10">
@@ -101,77 +62,7 @@ export default function PollPage() {
         </div>
       </div>
       <div className="container mx-auto px-4 py-8">
-        <Card className="w-full max-w-2xl mx-auto bg-white/80 backdrop-blur-sm border-0 shadow-xl">
-          <CardHeader className="pb-4">
-            <div className="flex items-start justify-between gap-4">
-              <div className="flex-1">
-                <CardTitle className="text-xl font-bold text-gray-900 leading-tight">
-                  {poll.title}
-                </CardTitle>
-                <p className="text-gray-600 mt-2 text-sm">{poll.description}</p>
-              </div>
-              <Badge className="bg-purple-100 text-purple-700 hover:bg-purple-200 transition-colors shrink-0">
-                <Users className="w-3 h-3 mr-1" />
-                {poll.totalVotes}
-              </Badge>
-            </div>
-          </CardHeader>
-          <CardContent className="space-y-3">
-            {poll.options.map((option: PollOption, index: number) => {
-              const percentage = getPercentage(option.votes, poll.totalVotes);
-              const isUserVote = voteStatus?.userVote === option.id;
-
-              return (
-                <div key={option.id} className="space-y-2">
-                  <div className="flex items-center justify-between">
-                    <span className="text-sm font-medium text-gray-700">{option.text}</span>
-                    <div className="flex items-center gap-2">
-                      <span className="text-xs text-gray-500">{option.votes.toLocaleString()} votes</span>
-                      <span className="text-sm font-semibold text-purple-600">{percentage}%</span>
-                      {isUserVote && <CheckCircle className="w-4 h-4 text-green-500" />}
-                    </div>
-                  </div>
-                  <div className="relative">
-                    <div className="w-full bg-gray-200 rounded-full h-3 overflow-hidden">
-                      <div
-                        className={`h-full rounded-full transition-all duration-1000 ease-out ${
-                          isUserVote ? "bg-gradient-to-r from-green-400 to-green-500" : "bg-gradient-to-r from-purple-400 to-purple-500"
-                        }`}
-                        style={{
-                          width: `${percentage}%`,
-                          animationDelay: `${index * 100}ms`,
-                        }}
-                      />
-                    </div>
-                  </div>
-                </div>
-              );
-            })}
-            {voteStatus?.hasVoted && (
-              <div className="mt-6 p-4 bg-green-50 rounded-xl border border-green-200">
-                <div className="flex items-center gap-2 text-green-700">
-                  <CheckCircle className="w-5 h-5" />
-                  <span className="font-medium">Thank you for voting!</span>
-                </div>
-                <p className="text-sm text-green-600 mt-1">Your vote has been recorded. Results are updated in real-time.</p>
-              </div>
-            )}
-            <div className="mt-4 flex gap-2">
-              <Link href="/" className="flex-1">
-                <Button className="w-full bg-gray-500 hover:bg-gray-600">Back to All Polls</Button>
-              </Link>
-              <ShareButton
-                buttonText="Share Poll"
-                cast={castConfig}
-                className="flex-1 bg-gradient-to-r from-green-500 to-green-600 hover:from-green-600 hover:to-green-700 text-white"
-                isLoading={false}
-              >
-                <Share2 className="w-4 h-4 mr-2" />
-                Share Poll
-              </ShareButton>
-            </div>
-          </CardContent>
-        </Card>
+        <PollCard poll={poll} userFid={userFid} showShareButton />
       </div>
       <div className="bg-white/50 backdrop-blur-sm border-t border-purple-100 mt-16">
         <div className="container mx-auto px-4 py-8 text-center">
